feat: 푸터에 다크모드 적용
